fix: ArraySlice<T> interface implementation for long indexing
- this[long index] indexer
- GetIndex/SetIndex with long index
- Slice(long start), Slice(long start, long length)
- Explicit IArraySlice implementations

Build has 439 cascading errors remaining across 50+ files.
Most are straightforward loop index changes (int → long).

Related to #584

+++ src/NumSharp.Core/Backends/Unmanaged/Interfaces/IArraySlice.cs
@@ -7,13 +7,13 @@ namespace NumSharp.Backends.Unmanaged
     {
         IMemoryBlock MemoryBlock { get; }
 
-        T GetIndex<T>(int index) where T : unmanaged;
+        T GetIndex<T>(long index) where T : unmanaged;
 
-        object GetIndex(int index);
+        object GetIndex(long index);
 
-        void SetIndex<T>(int index, T value) where T : unmanaged;
+        void SetIndex<T>(long index, T value) where T : unmanaged;
 
-        void SetIndex(int index, object value);
+        void SetIndex(long index, object value);
 
         new IArraySlice Clone();
 
@@ -23,7 +23,7 @@ namespace NumSharp.Backends.Unmanaged
 
         /// <param name="index"></param>
         /// <returns></returns>
-        unsafe object this[int index] {  get;  set; }
+        unsafe object this[long index] {  get;  set; }
 
         /// <summary>
         ///     Fills all indexes with <paramref name="value"/>.
@@ -36,7 +36,7 @@ namespace NumSharp.Backends.Unmanaged
         /// </summary>
         /// <param name="start">The index to start from</param>
         /// <remarks>Creates a slice without copying.</remarks>
-        IArraySlice Slice(int start);
+        IArraySlice Slice(long start);
 
         /// <summary>
         ///     Perform a slicing on this <see cref="IMemoryBlock"/> without copying data.
@@ -44,7 +44,7 @@ namespace NumSharp.Backends.Unmanaged
         /// <param name="start">The index to start from</param>
         /// <param name="count">The number of items to slice (not bytes)</param>
         /// <remarks>Creates a slice without copying.</remarks>
-        IArraySlice Slice(int start, int count);
+        IArraySlice Slice(long start, long count);
 
         /// <param name="destination"></param>
         void CopyTo<T>(Span<T> destination);
